added UserIngredient enpoints + readme update

+++ ingredient-service/src/main/java/com/chefathands/ingredient/repository/IngredientRepository.java
@@ -4,6 +4,6 @@ import com.chefathands.ingredient.model.Ingredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
     List<Ingredient> findByNameContainingIgnoreCase(String name);
 }
